Public interfaces

FunctionFusion.AbstractProvider — Type

AbstractProvider

Type common for all providers. Shall implement interface with methods

is_provider

inputs

outputs

short_description

provide

FunctionFusion.artifact_type — Method

artifact_type(Artifact)::Type

Get the Type associated with an Artifact

Example

@artifact A = Int
artifact_type(A) == Int

FunctionFusion.describe_provider — Method

describe_provider(x)::AbstractProvider

Returns Provider object describing the provider This function is only defined for providers, use is_provider(x) to check if x is a provider

FunctionFusion.inputs — Function

inputs(A::AbstractProvider)

Return all artifacts required by provider A

FunctionFusion.is_artifact — Method

is_artifact(::Type)

Check if a given type is an Artifact.

FunctionFusion.is_provider — Method

is_provider(x)::Boolean

Returns true if x is declared as a provider and false otherwise.

FunctionFusion.outputs — Function

outputs(A::AbstractProvider)

Returns all artifacts provided by provider A

FunctionFusion.provide — Function

provide(A::AbstractProvider, artifact::AbstractArtifact, context, parent::Union{AbstractProvider, Nothing})

Returns Expr which computes requested artifact using values from context for source

FunctionFusion.read_function_signature — Method

readfunctionsignature(func::Expr)::NamedTuple{(:name,:result,:arguments)}

Read the signature of a provider function and return a named tuple with the function's name, result type, and argument types.

The result can be a single artifact type or a tuple of artifact types, e.g.:

f(a::A)::B returns a single artifact

f(a::A)::(B, C) returns a tuple of artifacts

FunctionFusion.storage — Function

storage(A::AbstractProvider)

Returns storage container necessary for provider A

FunctionFusion.visualize — Method

visualize(what, [format]; parameters...)

Provides visualization in MIME format for what. what can be either provider or vector of providers If format is not specified then it is chosen automatically:

If GrahpViz package is not loaded then it is MIME"text/vnd.graphviz" which is a graphviz document.

If GraphViz package is loaded then it would be MIME"image/svg+xml" which is a .svg image.

Load GraphViz in your REPL to have visualization as picture (automatically displayed in VScode).

parameters:

mod=Main display names relative to that module

depth=Inf display only depth nesting of the graph

hide_types=False hide types of the artifacts

FunctionFusion.@algorithm — Macro

@algorithm name(Input,...)::Output = [Providers] @algorithm name(Input,...)::(Output1, Output2) = [Providers]

Generates function that implements the algorithm to compute given Outputs from Inputs using Providers. The output can be a single artifact or a tuple of artifacts.

FunctionFusion.@artifact — Macro

@artifact name[,name...] = Type

Declares Artifact associated with Type

Example

@artifact A = Int
@artifact B,C = Bool

FunctionFusion.@callback_provider — Macro

@callback_provider Name()::CallbackTypeName = ProviderName

Creates artifact CallbackTypeName which will point to argumentless function which will return result of ProviderName

FunctionFusion.@conditional — Macro

@conditional name::output_artifact = bool_artifact ? input_artifact1 : input_artifact2

Defines a ConditionalProvider that promotes one of the input_artifact to an output_artifact depending on a value of bool_artifact. The input_artifacts and output_artifact should be of the same type and bool_artifact shall be of type Bool

Example

@artifact A = Int
@artifact B = Int
@artifact C = Bool
@artifact D = Int

@conditional conditional::D = C ? A : B

FunctionFusion.@context — Macro

@context name = [artifacts_or_contexts...]

Creates structure name which contains set of artifacts and contexts. This structure implements single write, so it's elements can be stored only once. Access to elements of the structure is done via index operator [] where key is the Artifact or Context types

Example:

@artifact A,B Int
@context Ctx = [A, B]

ctx = Ctx()
isnothing(ctx[A]) == true

ctx[A] = 1
isnothing(ctx[A]) == false

FunctionFusion.@group — Macro

@group name = [Providers...]

Declares group of providers, such that all artifacts generated by providers are stored in that group. This can be used to group providers and give better names to set of artifacts of the group. All artifacts of the group would be available to the overlaying group / algorithm.

FunctionFusion.@invoke_with — Macro

@invoke_with Name = AlgorithmName{Substitutions}

FunctionFusion.@promote — Macro

@promote name(Input)::Output

Defines promote provider that assigns data from the Input Artifact to the Output Artifact. Both Artifacts have to share same data type.

FunctionFusion.@provider — Macro

1

@provider function name(arg::Artifact, ...)::Artifact ... end

2

@provider name(arg::Artifact, ...)::Artifact = x

2

@provider name(Artifact,...)::Artifact

3

@provider alias = name(Artifact, ...)::Artifact

Declares a provider with given inputs and output. All inputs + output must be unique artifacts.

3 versions of the syntax are supported: 1 and 2 - function definitions 3 - declare existing function as provider 4 - make an alias to existing function and declare it as a provider

FunctionFusion.@switch_provider — Macro

@switch_provider Name(Input)::Output = [Options...]

Creates switch provider that will map Input to switch which of Options include in the Output If Output is a AbstractVector or Tuple then Input had to be iteratable and it will construct it for as many Input elements as it has otherwise only one Artifact would be returned

FunctionFusion.@unimplemented — Macro

@unimplemented name(Artifact, ...)::Artifact

Declares an Unimplemented provider with given inputs and output. All inputs + output must be unique artifacts.